Began work on Run stage and cleaned up points
Began work on run stage, still not complete yet.

Main change was changing the way points were reported from stages.
before we were having the stage report a score and the test would
compile all the scores together. This system proved difficult to work
with.

Unified all the different point types into PointQuantity which can be
used to report points lost on a test, or how many points a specific
modifier is worth.

It was much easier to work with stages reporting points lost instead of
a score.

P.S. Sorry for how messy this commit is. I promise I'll do better.