Created Genos test executor
Created a test executor which will run the tests provided by users in
parallel and collate the results.

Genos should not panic on any system errors, but should always report
them to the student and fail the test. For setup tests, a system error
is considered fatal, and the autograder should exit after writing the
results. For normal tests, a system error should be output to the
student, but the rest of the tests should continue to run.