Make RichText Clone less expensive + Context setup
Add a fix to make cloning Richtext less expensive. It could contain very
large strings and we make frequent copies of the test results which
contain the test output which could contain rich text. Wrap it in an arc
to make cloning cheap.

Also added config for command line arguments.

Also added an app context. This will be used to wrap all things that are
required to run an instance of the autograder.